backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
alembic>=1.10.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
redis>=4.0.0
gunicorn>=21.2.0
pydantic>=2.0.0
python-multipart>=0.0.6